Fix class hw and attendance styling
