--- react-app/src/pages/ClassPage/AttendancePage/AttendancePage.module.css
@@ -1,145 +1,0 @@
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-}
-
-.topLevel {
-  margin-top: 1%;
-}
-
-.noteIcon {
-  width: 34px;
-  height: 24px;
-  background-color: var(--color-orange);
-  border-radius: 13px;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-left: 12px;
-  padding-right: 9px;
-  margin-left: 10%;
-}
-
-.dateSelection {
-  float: right;
-  width: 140px;
-  margin-right: 10%;
-  height: 25px;
-  font-size: large;
-}
-
-.inputs {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 80%;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 1%;
-  /* float: right; */
-}
-
-.box {
-  padding-top: 30px;
-  padding-bottom: 30px;
-  padding-left: 5px;
-  display: flex;
-  border-bottom: 1px solid #000000;
-}
-
-.bottomBox {
-  padding-bottom: 30px;
-  padding-top: 30px;
-  padding-left: 5px;
-  display: flex;
-}
-
-.boxTitle {
-  font-style: normal;
-  font-weight: 400;
-  font-size: 1.5rem;
-  width: 200px;
-  padding-left: 40px;
-  display: inline-block;
-}
-
-.selectAll {
-  width: 130px;
-  height: 40px;
-  float: right;
-  background: var(--color-orange);
-  border-radius: 30px;
-  font-size: 17px;
-  border: none;
-  margin-right: 10%;
-}
-
-.icons {
-  width: 75%;
-  display: flex;
-  justify-content: end;
-  margin-right: 20%;
-}
-
-.bottomLevel {
-  padding-top: 3%;
-}
-
-.statusLabel {
-  padding-left: 3%;
-}
-
-/* customized checkbox css*/
-.checkboxContainer {
-  display: inline-block;
-  position: relative;
-  padding-left: 25px;
-  margin-bottom: 15px;
-  cursor: pointer;
-}
-
-.checkboxContainer input {
-  position: absolute;
-  opacity: 0;
-  cursor: pointer;
-}
-
-.checkmark {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 20px;
-  width: 20px;
-  border: 3px solid black;
-  border-radius: 2px;
-}
-
-.checkboxContainer:hover input ~ .checkmark {
-  background-color: #ccc;
-}
-
-.checkboxContainer input:checked ~ .checkmark {
-  background-color: black;
-}
-
-.checkmark:after {
-  content: '';
-  position: absolute;
-  display: none;
-}
-
-.checkboxContainer input:checked ~ .checkmark:after {
-  display: flex;
-}
-
-.checkboxContainer .checkmark:after {
-  left: 7px;
-  top: 3px;
-  width: 5px;
-  height: 10px;
-  border: solid white;
-  border-width: 0 3px 3px 0;
-  transform: rotate(45deg);
-}
