+++ react-app/src/pages/ClassPage/ClassHomework/ClassHomework.module.css
@@ -1,0 +1,88 @@
+.mainContainer {
+  display: flex;
+  flex-direction: column;
+}
+
+.topLevel {
+  margin-top: 1%;
+}
+
+.noteButton {
+  margin-left: 10%;
+  border-radius: 13px;
+  border: none;
+  background-color: var(--color-orange);
+}
+
+.noteButton:hover {
+  cursor: pointer;
+  background-color: var(--color-yellow-orange);
+}
+
+.noteIcon {
+  width: 23px;
+  padding: 10px 12px 8px 12px;
+}
+
+.dateSelection {
+  float: right;
+  width: 140px;
+  margin-right: 10%;
+  height: 25px;
+  font-size: large;
+}
+
+.inputs {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  border: 1px solid black;
+  border-radius: 20px;
+  width: 80%;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 1%;
+}
+
+.box {
+  border-bottom: 1px solid black;
+  display: flex;
+  padding: 10px 40px 10px 10px;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.roundTop {
+  border-radius: none;
+}
+
+.roundBottom {
+  border: none;
+}
+
+.boxTitle {
+  padding-left: 40px;
+}
+
+.bottomLevel {
+  padding-top: 3%;
+  display: flex;
+  justify-content: center;
+  gap: 5%;
+}
+
+.bottomButton {
+  width: 150px;
+  height: 40px;
+  background: var(--color-orange);
+  color: white;
+  border-radius: 30px;
+  font-size: 1.05rem;
+  border: none;
+}
+
+.bottomButton:hover {
+  cursor: pointer;
+  background: var(--color-yellow-orange);
+}
