+++ react-app/src/pages/ClassPage/ClassStudents/ClassStudents.module.css
@@ -62,16 +62,19 @@
 }
 
 .certificateIcon:hover {
+  cursor: pointer;
   filter: invert(10%) sepia(83%) saturate(6663%) hue-rotate(253deg)
     brightness(76%) contrast(110%);
 }
 
 .profileIcon:hover {
+  cursor: pointer;
   filter: invert(72%) sepia(14%) saturate(6705%) hue-rotate(334deg)
     brightness(94%) contrast(89%);
 }
 
 .trashIcon:hover {
+  cursor: pointer;
   filter: invert(8%) sepia(100%) saturate(7043%) hue-rotate(352deg)
     brightness(87%) contrast(104%);
 }
