--- react-app/src/pages/ClassPage/HomeworkPage/HomeworkPage.module.css
@@ -1,114 +1,0 @@
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-}
-
-.topLevel {
-  margin-top: 1%;
-}
-
-.noteIcon {
-  width: 34px;
-  height: 24px;
-  background-color: var(--color-orange);
-  border-radius: 13px;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-left: 12px;
-  padding-right: 9px;
-  margin-left: 10%;
-}
-
-.dateSelection {
-  float: right;
-  width: 140px;
-  margin-right: 10%;
-  height: 25px;
-  font-size: large;
-}
-
-.inputs {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 80%;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 1%;
-  /* float: right; */
-}
-
-.box {
-  padding-top: 30px;
-  padding-bottom: 30px;
-  padding-left: 5px;
-  display: flex;
-  border-bottom: 1px solid #000000;
-}
-
-.bottomBox {
-  padding-bottom: 30px;
-  padding-top: 30px;
-  padding-left: 5px;
-  display: flex;
-}
-
-.boxTitle {
-  font-style: normal;
-  font-weight: 400;
-  font-size: 1.5rem;
-  width: 200px;
-  padding-left: 40px;
-  display: inline-block;
-}
-
-.bottomLevel {
-  padding-top: 3%;
-}
-
-.add {
-  width: 150px;
-  height: 40px;
-  float: right;
-  background: var(--color-orange);
-  border-radius: 30px;
-  font-size: 17px;
-  border: none;
-  margin-right: 5%;
-}
-
-.remove {
-  width: 150px;
-  height: 40px;
-  float: right;
-  background: var(--color-orange);
-  border-radius: 30px;
-  font-size: 17px;
-  border: none;
-  margin-right: 5%;
-}
-
-.selectAll {
-  width: 150px;
-  height: 40px;
-  float: right;
-  background: var(--color-orange);
-  border-radius: 30px;
-  font-size: 17px;
-  border: none;
-  margin-right: 5%;
-}
-
-.save {
-  width: 150px;
-  height: 40px;
-  float: right;
-  background: var(--color-orange);
-  border-radius: 30px;
-  font-size: 17px;
-  border: none;
-  margin-right: 19%;
-}
