+++ react-app/src/pages/LoginPage/ForgotPasswordModal/ForgotPassword.module.css
@@ -87,5 +87,6 @@
 }
 
 .resetButton:hover {
+  cursor: pointer;
   background: var(--color-yellow-orange);
 }
