+++ react-app/src/pages/SettingsPage/ResetPassword/ResetPassword.module.css
@@ -95,5 +95,6 @@ form {
 }
 
 .resetButton:hover {
+  cursor: pointer;
   background: var(--color-yellow-orange);
 }
