+++ react-app/src/pages/StudentRosterPage/StudentList/StudentList.module.css
@@ -43,11 +43,13 @@
 }
 
 .profileIcon:hover {
+  cursor: pointer;
   filter: invert(72%) sepia(14%) saturate(6705%) hue-rotate(334deg)
     brightness(94%) contrast(89%);
 }
 
 .trashIcon:hover {
+  cursor: pointer;
   filter: invert(8%) sepia(100%) saturate(7043%) hue-rotate(352deg)
     brightness(87%) contrast(104%);
 }
